If TOR_CONFIGURE_ONLY=1, only configure Tor.
This option prevents Tor Launcher from starting a Tor process but it
will try to configure any already started Tor process without taking
ownership of it. This option is intended for situations where Tor
Launcher is supposed to configure an already started, system-wide Tor
instance.

Note that it's not the same as TOR_SKIP_LAUNCH (and, equivalently,
`extensions.torlauncher.start_tor`), which prevents *both* start and
configuration of Tor.