Bug #10418: Provide "Use Default Bridges" option.
To enable this feature, add one or more preference values like:
  extensions.torlauncher.default_bridge.TYPE.1
  extensions.torlauncher.default_bridge.TYPE.2
Other new preferences:
  extensions.torlauncher.default_bridge_type (stores chosen type)
  extensions.torlauncher.default_bridge_recommended_type
The text "(recommended)" is displayed next to the recommended type, and
it is selected by default.  The default bridges are reconfigured each
time the browser starts up; if no default bridges of the configured type
are available, an error message is displayed and the user is placed on
the bridge settings panel within the network configuration wizard.