Improve how we import, bundle and register translations.
Previously chrome.manifest defined locales for pretty much every
existing locale. Many of these were out-of-date (lacking entities,
which results in fatal XUL errors) or completely absent (also fatal).

Instead of maintaining a static list of which locales to register we
generate it programatically, making sure that only those that seem
complete are included. We also make sure to register "en-US" since
it's the hardcoded fallback for unregistered locales. Note that it's
currently called "en" in the Transifex translation repos.

The BUNDLE_LOCALES environment variable can be set to specify a list
of which locales to bundle ("en-US" is always implied). If it's empty
we optimistically try to include all locales. For this reason
import-translations.sh imports all translations from Transifex, and
it's important to run it before packaging since up-to-date
translations are necessary to avoid the issues mentioned above.