Bug 20429: Do not open progress window if TOR_SKIP_LAUNCH=1
After saving tor settings, if TOR_SKIP_LAUNCH=1 or if
extensions.torlauncher.start_tor is false, avoid opening the progress
dialog to monitor bootstrap progress. This avoids displaying a progress
window that will not make progress (and the situation was made worse by
the fact that clicking Cancel set DisableNetwork=1).

Also, leave the network settings dialog open if we fail to set
DisableNetwork=0.