Bug 20761: Tor Browser 6.5a4 is ignoring additional SocksPorts
When starting tor, pass "+__ControlPort" and "+__SocksPort" in the
args. This causes Tor Launcher's control and SOCKS ports to be added
to any ports the user has defined in their torrc and also prevents
the ports defined by Tor Launcher from being written to torrc.

Also, perform a one-time removal of the torrc lines that conflict with
the ports that Tor Launcher will use when starting tor.

This change depends on the fix for #20956.